+++ src/tools/pes/pesgeneral.f90
@@ -7,21 +7,22 @@ module pesgeneral
   use constants
   use formulas_mod
   use input_params_mod
+  use iso_fortran_env, only: real64
   use general_utils
   use general_vars
-  
   implicit none
-  real*8, parameter :: r0 = 2.2819d0
+
+  real(real64), parameter :: r0 = 2.2819d0
   ! Potentials on 3D grid
-  real*8, allocatable :: potvib(:,:,:) ! Vibrational
-  real*8, allocatable :: pottot(:,:,:) ! Total
+  real(real64), allocatable :: potvib(:,:,:) ! Vibrational
+  real(real64), allocatable :: pottot(:,:,:) ! Total
   ! Rotation
-  real*8 jx,jy,jz    ! Components of total angular momentum J
+  real(real64) jx,jy,jz    ! Components of total angular momentum J
   integer jlarge     ! Absolute value of J
   integer klarge     ! Projection onto Z axis
   ! Miscellaneous
-  real*8 shift       ! Used for shifting PES down
-  real*8 threshold   ! Threshold in the lowest channel
+  real(real64) shift       ! Used for shifting PES down
+  real(real64) threshold   ! Threshold in the lowest channel
 
 contains
   
@@ -33,7 +34,7 @@ contains
   subroutine init_pots_general(params)
     class(input_params), intent(in) :: params
     integer :: a1, a2, a3, total_code
-    real*8 maxmu
+    real(real64) maxmu
 
     ! setup global variables
     mol = str2int(params % molecule)
@@ -75,9 +76,9 @@ contains
   !  Calculates rotational potential in symmetric top approximation.
   !-----------------------------------------------------------------------
   function calc_potrot(rho, theta, J, K) result(res)
-    real*8 :: rho, theta
+    real(real64) :: rho, theta
     integer, optional, intent(in) :: J, K
-    real*8 :: res
+    real(real64) :: res
     res = calc_potrotj2(rho, theta, J) + calc_potrotk2(rho, theta, K)
   end function
 
@@ -85,11 +86,11 @@ contains
   !  Calculates J**2 component of rotational potential.
   !-----------------------------------------------------------------------
   function calc_potrotj2(rho, theta, J) result(res)
-    real*8 :: rho, theta
+    real(real64) :: rho, theta
     integer, optional, intent(in) :: J
-    real*8 :: res
+    real(real64) :: res
     integer :: J_act
-    real*8 :: a, b
+    real(real64) :: a, b
 
     a = get_rotational_a(mu, rho, theta)
     b = get_rotational_b(mu, rho, theta)
@@ -101,11 +102,11 @@ contains
   !  Calculates K**2 component of rotational potential.
   !-----------------------------------------------------------------------
   function calc_potrotk2(rho, theta, K) result(res)
-    real*8 :: rho, theta
+    real(real64) :: rho, theta
     integer, optional, intent(in) :: K
-    real*8 :: res
+    real(real64) :: res
     integer :: K_act
-    real*8 :: a, b, c
+    real(real64) :: a, b, c
 
     a = get_rotational_a(mu, rho, theta)
     b = get_rotational_b(mu, rho, theta)
@@ -117,9 +118,9 @@ contains
   !-----------------------------------------------------------------------
   !  Calculates asymmetric component of rotational potential.
   !-----------------------------------------------------------------------
-  real*8 function calc_potrotasym(rho,tet)
-    real*8 rho, tet
-    real*8 :: a, b
+  real(real64) function calc_potrotasym(rho,tet)
+    real(real64) rho, tet
+    real(real64) :: a, b
     a = get_rotational_a(mu, rho, tet)
     b = get_rotational_b(mu, rho, tet)
     calc_potrotasym = (a - b) / 2
@@ -128,16 +129,16 @@ contains
   !-----------------------------------------------------------------------
   !  Calculates rotational potential at a given point.
   !-----------------------------------------------------------------------
-  real*8 function calc_potrotfull(rho,tet)
-    real*8 rho,tet
+  real(real64) function calc_potrotfull(rho,tet)
+    real(real64) rho,tet
     calc_potrotfull = ( jx**2/(1-sin(tet)) + jy**2/(1+sin(tet)) + jz**2/(2*sin(tet)**2) ) / (mu*rho**2)
   end function
 
   !-----------------------------------------------------------------------
   !  Calculates extra potential term at a given point.
   !-----------------------------------------------------------------------
-  real*8 function calc_potxtr(rho,tet)
-    real*8 rho,tet
+  real(real64) function calc_potxtr(rho,tet)
+    real(real64) rho,tet
     calc_potxtr = -(0.25d0+4/(sin(2*tet)**2))/(2d0*mu*rho**2)
   end function
