+++ src/utils/general/general_real.F90
@@ -2,14 +2,15 @@ module general_real_mod
 #include "type_list.macro"
 #define TYPE_ID REAL_ID
 #include "type_attributes.macro"
-use general_char_str_mod
+  use general_char_str_mod
+  use iso_fortran_env, only: real64
 #include "general_template.F90"
 
 !-------------------------------------------------------------------------------------------------------------------------------------------
-! Converts real*8 to string
+! Converts real(real64) to string
 !-------------------------------------------------------------------------------------------------------------------------------------------
   function num2str_real(num, format) result(str)
-    real*8, intent(in) :: num
+    real(real64), intent(in) :: num
     character(*), intent(in), optional :: format
     character(:), allocatable :: str
     character(256) :: buffer
@@ -21,11 +22,11 @@ use general_char_str_mod
   end function
 
 !-------------------------------------------------------------------------------------------------------------------------------------------
-! Converts string to real*8
+! Converts string to real(real64)
 !-------------------------------------------------------------------------------------------------------------------------------------------
   elemental function str2real(str) result(real_num)
     character(*), intent(in) :: str
-    real*8 :: real_num
+    real(real64) :: real_num
     read(str, *) real_num
   end function
